Fix narrowing UB, add comments

+++ include/co_usb/tfer/endpoint.hpp
@@ -40,11 +40,11 @@ template <ep_direction Direction> struct endpoint
     {
         if constexpr (Direction == ep_direction::out)
         {
-            return {ep ^ LIBUSB_ENDPOINT_IN, iface.dev()};
+            return {static_cast<uint8_t>(ep ^ LIBUSB_ENDPOINT_IN), iface.dev()};
         }
         else if constexpr (Direction == ep_direction::in)
         {
-            return {ep | LIBUSB_ENDPOINT_IN, iface.dev()};
+            return {static_cast<uint8_t>(ep | LIBUSB_ENDPOINT_IN), iface.dev()};
         }
         return {ep, iface.dev()};
     }
